:sparkles: 이벤트 이미지 등록
이벤트 이미지 등록
