+++ Front-end/emart-fresh-app/src/pages/myPageMenuBar/StoreManager.tsx
@@ -6,10 +6,16 @@ export default function StoreManager() {
   return (
     <div>
       <div className={styles.mypagemenubar}>
-        <Link to="/mypageMain">개인정보관리</Link>
+        <Link to="/">상품등록</Link>
+      </div>
+      <div className={styles.mypagemenubar}>
+        <Link to="/">발주현황확인</Link>
       </div>
       <div className={styles.mypagemenubar}>
         <Link to="/order-request">발주 보내는 페이지</Link>
+      </div>
+      <div className={styles.mypagemenubar}>
+        <Link to="/">매출현황</Link>
       </div>
     </div>
   );
